+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/interop/generic/ReadArrayElementNode.java
@@ -9,7 +9,7 @@ import org.enso.interpreter.Constants;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.epb.node.CoercePrimitiveNode;
 import org.enso.interpreter.node.expression.foreign.CoerceNothing;
-import org.enso.interpreter.runtime.Context;
+import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.builtin.Builtins;
 import org.enso.interpreter.runtime.error.PanicException;
 
@@ -31,12 +31,12 @@ public class ReadArrayElementNode extends Node {
       return nothingCoercion.execute(coercion.execute(library.readArrayElement(array, index)));
     } catch (UnsupportedMessageException e) {
       err.enter();
-      Builtins builtins = Context.get(this).getBuiltins();
+      Builtins builtins = EnsoContext.get(this).getBuiltins();
       throw new PanicException(
           builtins.error().makeTypeError(builtins.array(), array, "array"), this);
     } catch (InvalidArrayIndexException e) {
       err.enter();
-      Builtins builtins = Context.get(this).getBuiltins();
+      Builtins builtins = EnsoContext.get(this).getBuiltins();
       throw new PanicException(builtins.error().makeInvalidArrayIndexError(array, index), this);
     }
   }